pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "fpdf>=1.7.2",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pymupdf>=1.25.5",
    "pypandoc>=1.15",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "python-pptx>=1.0.2",
    "xlrd>=2.0.1",
]
